Create parent folders once for all stacks and syncs
There is an eventual consistency behaviour in how grafana
handles and creates dashboard folders. In particular for "top level folders"
This is problematic as it leads to, sometimes, duplicating the toplevel folders
in particular when there are multiple synchronised folders.

To work-around this, create the parent folder once for all in all stacks
at the begining of the Publishing

Also refactor the retry mechanism as this change highlights difficulties
apparently when the parent folder is created with a different client at the begining.
In those cases we may see errors like `[GET /folders][403] getFoldersForbidden {"message":"Access denied"}`

To work around this, use an exponential backoff to retry failed stacks

Change-Id: Ie6a974ac083af52326066dc28b666baafb8cdc47